pyproject.toml
[tool.poetry]
name = "my-rel-gen"
version = "0.1.0"
description = ""
authors = ["Bryan Hu <bryan.hu.2020@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
toml = "^0.10.2"
PyGithub = "^1.54.1"
halo = "^0.0.31"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.812"
pylint = "^2.7.2"
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
